requirements.txt
uvicorn
fastapi
requests
python-jose
pyjwt